Rename tonal module to core
