+++ package.json
@@ -14,8 +14,8 @@
     "build:browser": "cd packages/modules && rollup -c=rollup.browser.config.js",
     "build:clean": "lerna exec --parallel -- rimraf dist/",
     "test": "jest --coverage",
-    "test:lint": "tslint --fix --project tsconfig.json -t codeFrame 'packages/**/*.ts'",
-    "test:ci": "yarn test:lint && yarn build && yarn test -- --no-cache"
+    "lint": "tslint --fix --project tsconfig.json -t codeFrame 'packages/**/*.ts'",
+    "test:ci": "yarn lint && yarn build && yarn test -- --no-cache"
   },
   "husky": {
     "hooks": {
