+++ packages/abc-notation/package.json
@@ -17,7 +17,7 @@
   ],
   "types": "dist/abc-notation/index.d.ts",
   "dependencies": {
-    "@tonaljs/tonal": "^3.2.2"
+    "@tonaljs/core": "^3.2.2"
   },
   "author": "danigb@gmail.com",
   "license": "MIT",
