+++ packages/array/package.json
@@ -6,7 +6,7 @@
     "array",
     "music",
     "theory",
-    "@tonaljs/tonal"
+    "@tonaljs/core"
   ],
   "main": "dist/index.es5.js",
   "module": "dist/index.esnext.js",
@@ -15,7 +15,7 @@
   ],
   "types": "dist/array/index.d.ts",
   "dependencies": {
-    "@tonaljs/tonal": "^3.2.2"
+    "@tonaljs/core": "^3.2.2"
   },
   "author": "danigb@gmail.com",
   "license": "MIT",
