+++ packages/chord-dictionary/package.json
@@ -8,7 +8,7 @@
     "chords",
     "music",
     "theory",
-    "@tonaljs/tonal"
+    "@tonaljs/core"
   ],
   "main": "dist/index.es5.js",
   "module": "dist/index.esnext.js",
