+++ packages/core/src/interval.ts
@@ -89,7 +89,7 @@ const cache: { [key in string]: Interval | NoInterval } = {};
  * @return {Object} the interval properties
  *
  * @example
- * import { interval } from '@tonaljs/tonal'
+ * import { interval } from '@tonaljs/core'
  * interval('P5').semitones // => 7
  * interval('m3').type // => 'majorable'
  */
