+++ packages/core/src/note.ts
