+++ packages/key/index.ts
@@ -1,6 +1,6 @@
+import { accToAlt, altToAcc, note, transpose } from "@tonaljs/core";
 import { transposeFifths } from "@tonaljs/note";
 import { romanNumeral } from "@tonaljs/roman-numeral";
-import { accToAlt, altToAcc, note, transpose } from "@tonaljs/tonal";
 
 export interface Key {
   type: "major" | "minor";
