+++ packages/mode/package.json
@@ -6,7 +6,7 @@
     "mode",
     "music",
     "theory",
-    "@tonaljs/tonal"
+    "@tonaljs/core"
   ],
   "main": "dist/index.es5.js",
   "module": "dist/index.esnext.js",
@@ -16,7 +16,7 @@
   "types": "dist/mode/index.d.ts",
   "dependencies": {
     "@tonaljs/pcset": "^3.2.3",
-    "@tonaljs/tonal": "^3.2.2"
+    "@tonaljs/core": "^3.2.2"
   },
   "author": "danigb@gmail.com",
   "license": "MIT",
