+++ packages/modules/index.ts
@@ -2,6 +2,7 @@ import * as AbcNotation from "@tonaljs/abc-notation";
 import * as Array from "@tonaljs/array";
 import * as Chord from "@tonaljs/chord";
 import * as ChordDictionary from "@tonaljs/chord-dictionary";
+import * as Tonal from "@tonaljs/core";
 import * as Interval from "@tonaljs/interval";
 import * as Key from "@tonaljs/key";
 import * as Midi from "@tonaljs/midi";
@@ -13,7 +14,6 @@ import * as Range from "@tonaljs/range";
 import * as RomanNumeral from "@tonaljs/roman-numeral";
 import * as Scale from "@tonaljs/scale";
 import * as ScaleDictionary from "@tonaljs/scale-dictionary";
-import * as Tonal from "@tonaljs/tonal";
 
 export {
   AbcNotation,
