+++ packages/modules/package.json
@@ -32,7 +32,7 @@
     "@tonaljs/roman-numeral": "^3.3.1",
     "@tonaljs/scale": "^3.3.4",
     "@tonaljs/scale-dictionary": "^3.4.2",
-    "@tonaljs/tonal": "^3.2.2"
+    "@tonaljs/core": "^3.2.2"
   },
   "author": "danigb@gmail.com",
   "license": "MIT",
