+++ packages/note/package.json
@@ -6,7 +6,7 @@
     "note",
     "music",
     "theory",
-    "@tonaljs/tonal"
+    "@tonaljs/core"
   ],
   "main": "dist/index.es5.js",
   "module": "dist/index.esnext.js",
@@ -16,7 +16,7 @@
   "types": "dist/note/index.d.ts",
   "dependencies": {
     "@tonaljs/midi": "^3.3.1",
-    "@tonaljs/tonal": "^3.2.2"
+    "@tonaljs/core": "^3.2.2"
   },
   "author": "danigb@gmail.com",
   "license": "MIT",
