+++ packages/pcset/package.json
@@ -7,7 +7,7 @@
     "midi",
     "music",
     "theory",
-    "@tonaljs/tonal"
+    "@tonaljs/core"
   ],
   "main": "dist/index.es5.js",
   "module": "dist/index.esnext.js",
@@ -17,7 +17,7 @@
   "types": "dist/pcset/index.d.ts",
   "dependencies": {
     "@tonaljs/array": "^3.2.3",
-    "@tonaljs/tonal": "^3.2.2"
+    "@tonaljs/core": "^3.2.2"
   },
   "author": "danigb@gmail.com",
   "license": "MIT",
