+++ packages/progression/package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@tonaljs/chord": "^3.4.0",
     "@tonaljs/roman-numeral": "^3.3.1",
-    "@tonaljs/tonal": "^3.2.2"
+    "@tonaljs/core": "^3.2.2"
   },
   "author": "danigb@gmail.com",
   "license": "MIT",
