+++ packages/scale-dictionary/package.json
@@ -7,7 +7,7 @@
     "dictionary",
     "music",
     "theory",
-    "@tonaljs/tonal"
+    "@tonaljs/core"
   ],
   "main": "dist/index.es5.js",
   "module": "dist/index.esnext.js",
