+++ packages/scale/package.json
@@ -20,7 +20,7 @@
     "@tonaljs/chord-dictionary": "^3.4.0",
     "@tonaljs/pcset": "^3.2.3",
     "@tonaljs/scale-dictionary": "^3.4.2",
-    "@tonaljs/tonal": "^3.2.2"
+    "@tonaljs/core": "^3.2.2"
   },
   "author": "danigb@gmail.com",
   "license": "MIT",
